Expand Card Class, Fix various omissions, Exercise via example file.
This commit fixes a couple toArray() issues that were present in the
previous commit. It also fixes an issue in the addition to the Message
class that allows the user to set ‘message_format’ to non-“html” which
is invalid. Constructors are added to the helper classes for the
required fields.

An examples directory is created with cards.php which exhibits the
various demonstration cards from the API documentation.